% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combineCLogitMI.R
\name{combineCLogitMI}
\alias{combineCLogitMI}
\title{Combine conditional logit model results fitted to multiply imputed datasets}
\usage{
combineCLogitMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of conditional logit fitted model outputs from the survival package}
}
\description{
This sltools function allows you to combine multiple imputation conditional logit model results from the survival package, extract results for the texreg package, and also output summary tables.
}

