% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{combineHurdleMI}
\alias{combineHurdleMI}
\title{A sltools function to combine hurdle model results fitted to multiply imputed datasets}
\usage{
combineHurdleMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of hurdle fitted model outputs from the pscl package}
}
\description{
This function allows you to combine MI hurdle results from the pscl package, extract results for the texreg package, and also put into a data frame.
}

