% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combineLmerMI.R
\name{combineLmerMI}
\alias{combineLmerMI}
\title{Combine lmer model results fitted to multiply imputed datasets}
\usage{
combineLmerMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of lmer fitted model outputs from the lme4 package}
}
\description{
This sltools function allows you to combine lmer results from the lme4 package, extract results for the texreg package, and also output a summary table.
}

