% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{combineZIcountMI}
\alias{combineZIcountMI}
\title{A sltools function to combine zero-inflated count model results fitted to multiply imputed datasets}
\usage{
combineZIcountMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of hurdle fitted model outputs from the pscl package}
}
\description{
This function allows you to combine MI zero-inflated results from the pscl package, extract results for the texreg package, and also put into a data frame.
}

