% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeLogitEV.R
\name{computeLogitEV}
\alias{computeLogitEV}
\title{Simulate quantities of interest for logit models}
\usage{
computeLogitEV(data = NULL, x = NULL, coef = NULL, vcov = NULL,
  num = NULL)
}
\arguments{
\item{data}{imputed datasets in amelia object}

\item{x}{a vector of values for predictors}

\item{coef}{a matrix of coefficients}

\item{vcov}{a list of variance covariance matrices}

\item{num}{number of simulations}
}
\description{
This function follows Zelig's simulation method and computes expected and predicted values for logit models fitted to multiply imputed datasets.
}

