% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{computeNegBinEV}
\alias{computeNegBinEV}
\title{A sltools function to simulate quantities of interest for negative binomial models}
\usage{
computeNegBinEV(data = NULL, x = NULL, coef = NULL, vcov = NULL,
  num = NULL, theta = NULL)
}
\arguments{
\item{data}{imputed datasets in amelia object}

\item{x}{a vector of values for predictors}

\item{coef}{a matrix of coefficients}

\item{vcov}{a list of variance covariance matrices}

\item{num}{number of simulations}

\item{theta}{a vector of theta parameters from each fitted negative binomial model results}
}
\description{
This function follows Zelig's simulation method and computes expected and predicted values for negative binomial models fitted to multiply imputed datasets.
}

