% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qiLogit}
\alias{qiLogit}
\title{A sltools function to simulate first differences for logit models}
\usage{
qiLogit(data = NULL, x = NULL, x1 = NULL, coef = NULL, vcov = NULL,
  num = 10000)
}
\arguments{
\item{data}{imputed datasets in amelia object}

\item{x}{a vector of values for predictors}

\item{x1}{a second vector of values for predictors}

\item{coef}{a matrix of coefficients}

\item{vcov}{a list of variance covariance matrices}

\item{num}{number of simulations. Default = 10000.}
}
\description{
This function uses the \code{computeLogitEV} function to simulate expected values, predicted values, and first differences for logit models.
}

