% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qiNegBin}
\alias{qiNegBin}
\title{Simulate first differences for negative binomial models}
\usage{
qiNegBin(data = NULL, x = NULL, x1 = NULL, coef = NULL, vcov = NULL,
  num = 10000, theta = NULL)
}
\arguments{
\item{data}{imputed datasets in amelia object}

\item{x}{a vector of values for predictors}

\item{x1}{a second vector of values for predictors}

\item{coef}{a matrix of coefficients}

\item{vcov}{a list of variance covariance matrices}

\item{num}{number of simulations. Default = 10000.}

\item{theta}{a vector of theta parameters from each fitted negative binomial model results}
}
\description{
This function uses the \code{computeNegBinEV} function to simulate expected values, predicted values, and first differences for negative binomial models.
}

