% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simHurdle.R
\name{simHurdle}
\alias{simHurdle}
\title{Simulate quantities of interest for hurdle models}
\usage{
simHurdle(data = NULL, x = NULL, x1 = NULL, hurdle.out = NULL,
  num = NULL)
}
\arguments{
\item{data}{imputed datasets in amelia object}

\item{x}{a vector of values for predictors}

\item{x1}{a second vector of values for predictors}

\item{hurdle.out}{a list of hurdle model outputs}

\item{num}{number of simulations}
}
\description{
This sltools function simulates quantities of interests from a list of hurdle model outputs fitted to Amelia MI objects.
}

