% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/texregLmRobust.R
\docType{class}
\name{texregLmRobust}
\alias{texregLmRobust}
\title{texreg extracting function for Zelig linear models fitted with robust SEs}
\arguments{
\item{names}{vector of variable names}

\item{coef}{vector of coefficients}

\item{se}{vector of standard errors}

\item{pval}{vector of p-values}

\item{n}{scalar of sample size}
}
\description{
This function extracts parameters from Zelig 3.5.5 lm objects with robust SEs and stores them in a new texreg object called "texregLmRobust," which feeds into the texreg package.
}

