% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{texregMIrelogit}
\alias{texregMIrelogit}
\title{A sltools function to extract and create a new texreg class}
\arguments{
\item{names}{vector of variable names}

\item{coef}{vector of coefficients}

\item{se}{vector of standard errors}

\item{pval}{vector of p-values}

\item{n}{scalar of sample size}

\item{aic}{scalar of AIC}
}
\description{
This function extracts parameters from multiple imputation relogit objects and stores them in a new texreg object called "texregMIrelogit," which feeds into the texreg package.
}

