% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{coefficientPlot}
\alias{coefficientPlot}
\title{A sltools coefficient plotting function}
\usage{
coefficientPlot(models, alpha = 0.05, model.names = "", var.order = "",
  var.names = NULL, omit.intercept = TRUE, omit.factor = TRUE)
}
\arguments{
\item{models}{a list of models}

\item{alpha}{scalar of alpha}

\item{model.names}{list of model names}

\item{var.order}{vector of variable ordering}

\item{var.names}{list of variable names}

\item{omit.intercept}{whether to omit intercept, default = TRUE}

\item{omit.factor}{whether to omit factors, default = TRUE}
}
\description{
This function allows you to extract coefficients from a list of models and create side-by-side coefplots.
}

