% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combineGlmMI}
\alias{combineGlmMI}
\title{Combine glm model results fitted to multiply imputed datasets}
\usage{
combineGlmMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of glm fitted model outputs from the pscl package}
}
\description{
This sltools function allows you to combine multiple imputation glm model results from the stats package, extract results for the texreg package, and also output summary tables.
}

