% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{combineGlmMI}
\alias{combineGlmMI}
\title{A sltools function to combine glm model results fitted to multiply imputed datasets}
\usage{
combineGlmMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of glm fitted model outputs from the pscl package}
}
\description{
This function allows you to combine glm results from the stats package, extract results for the texreg package, and also put into a data frame.
}

