% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combineHurdleMI}
\alias{combineHurdleMI}
\title{Combine hurdle model results fitted to multiply imputed datasets}
\usage{
combineHurdleMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of hurdle fitted model outputs from the pscl package. For example, llply(data.mi$imputations, function(x) return(hurdle(y ~ x, dist = "negbin", zero.dist = "binomial", link = "logit", data = x)}
}
\description{
This sltools function allows you to combine multiple imputation hurdle model results from the pscl package, extract results for the texreg package, and also output summary tables.
}

