% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combineLmerMI}
\alias{combineLmerMI}
\title{Combine lmer model results fitted to multiply imputed datasets}
\usage{
combineLmerMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of lmer fitted model outputs from the lme4 package}
}
\description{
This sltools function allows you to combine lmer results from the lme4 package, extract results for the texreg package, and also output a summary table.
}

