% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{combineLmerMI}
\alias{combineLmerMI}
\title{A sltools function to combine lmer model results fitted to multiply imputed datasets}
\usage{
combineLmerMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of glm fitted model outputs from the pscl package}
}
\description{
This function allows you to combine lmer results from the lme4 package, extract results for the texreg package, and also put into a data frame.
}

