% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combineZIcountMI.R
\name{combineZIcountMI}
\alias{combineZIcountMI}
\title{Combine zero-inflated count model results fitted to multiply imputed datasets}
\usage{
combineZIcountMI(fitted.obj = NULL)
}
\arguments{
\item{fitted.obj}{a list of hurdle fitted model outputs from the pscl package}
}
\description{
This sltools function allows you to combine multiple imputation zero-inflated model results from the pscl package, extract results for the texreg package, and also output summary tables.#'
}

