% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{texregMI}
\alias{texregMI}
\title{texreg extracting function for models fitted to multiply imputed data}
\arguments{
\item{names}{vector of variable names}

\item{coef}{vector of coefficients}

\item{se}{vector of standard errors}

\item{pval}{vector of p-values}

\item{n}{scalar of sample size}

\item{loglik}{scalar of log-likelihood}

\item{aic}{scalar of AIC}
}
\description{
This function extracts parameters from multiple imputation objects and stores them in a new texreg object called "texregMI," which feeds into the texreg package.
}

